requirements.txt
fastapi==0.68.1
click==8.0.3
sqlalchemy==1.4.23
pydantic==1.8.2
python-binance==1.0.15
python-telegram-bot==13.7
pyyaml==5.4.1
redis==4.0.2
pika==1.2.0
cryptography==3.4.8
uvicorn==0.15.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0 